implemented more instructions; added op descriptions for logging
I've been trying to run Adjustris to chase out instructions I haven't implemented yet, and figure out if the GPU implementation works. It crashes, but it takes a while on my laptop. Eventually it tries to write to an invalid address.

I added descriptions to each operation so I can print out the instruction as it executes, like `add a, d8`, but eventually would like to print what the fetched immediate actually was.