Fix failing GPU tests
Reworked stat register implementation so that updating GPU's mode, ly,
and lyc members will update the stat register appropriately. This is an
optimization for ROMs that poll the stat register--this way its value
doesn't have to be constructed from the state of the GPU, but the
implementation is slightly more complicated as a result.